Moved classes into separate files and renamed with PascalCase format
