+++ src/GameEngine.java
@@ -1,0 +1,22 @@
+import javax.swing.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+public class  GameEngine implements ActionListener {
+    Game parent;
+
+    GameEngine(Game parent) {
+        this.parent = parent;
+    }
+
+    @Override
+    public void actionPerformed(ActionEvent e) {
+        Object eventSource = e.getSource();
+        JButton clickedButton = (JButton) eventSource;
+        String name = clickedButton.getName();
+        String[] xy = clickedButton.getName().split(" ", 2);
+        int x = Integer.parseInt(xy[0]);
+        int y = Integer.parseInt(xy[1]);
+        parent.buttonClicked(x, y);
+    }
+}
