+++ src/Minesweeper.java
@@ -1,14 +1,12 @@
 public class Minesweeper {
-    private static Minesweeper minesweeper;
-    private static game newGame;
 
     public void proceed(int size) {
-        newGame = new game(size);
+        Game newGame = new Game(size);
         newGame.main(newGame, size);
     }
 
     public static void main(String[] args) {
-        minesweeper = new Minesweeper();
+        Minesweeper minesweeper = new Minesweeper();
         minesweeper.proceed(10);
     }
 }
