+++ src/MyMouseListener.java
@@ -1,0 +1,31 @@
+import java.awt.event.MouseEvent;
+import java.awt.event.MouseListener;
+
+public class MyMouseListener implements MouseListener {
+    Game parent;
+
+    MyMouseListener(Game parent) {
+        this.parent = parent;
+    }
+
+    public void mouseExited(MouseEvent arg0){
+    }
+    public void mouseEntered(MouseEvent arg0){
+    }
+    public void mousePressed(MouseEvent arg0){
+    }
+    public void mouseClicked(MouseEvent arg0){
+    }
+
+    @Override
+    public void mouseReleased(MouseEvent arg0) {
+        /*
+        if(SwingUtilities.isRightMouseButton(arg0)){
+            Object eventSource = arg0.getSource();
+            JButton clickedButton = (JButton) eventSource;
+            String[] xy = clickedButton.getName().split(" ", 2);
+            int x = Integer.parseInt(xy[0]);
+            int y = Integer.parseInt(xy[1]);
+        }*/
+    }
+}
